Cleanup: move unused translation helpers to migrations
These helpers are no longer used in the main source tree.

changelog_begin
changelog_end
