+++ ledger/sandbox/src/main/scala/db/migration/postgres/V25__Backfill_Participant_Events.scala
@@ -6,8 +6,8 @@ package db.migration.postgres
 import com.daml.ledger.participant.state.v1.Offset
 import com.daml.lf.data.Ref
 import com.daml.platform.events.EventIdFormatter
-import com.daml.platform.store.serialization.TransactionSerializer
 import db.migration.postgres.v25_backfill_participant_events.V25TransactionsWriter
+import db.migration.translation.TransactionSerializer
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
 
 class V25__Backfill_Participant_Events extends BaseJavaMigration {
