+++ ledger/sandbox/src/main/scala/db/migration/translation/ContractSerializer.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store.serialization
+package db.migration.translation
 
 import java.io.InputStream
