+++ ledger/sandbox/src/main/scala/db/migration/translation/TransactionSerializer.scala
@@ -1,16 +1,16 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store.serialization
+package db.migration.translation
 
 import java.io.InputStream
 
+import com.daml.ledger.EventId
 import com.daml.lf.archive.{Decode, Reader}
 import com.daml.lf.transaction._
 import com.daml.lf.value.Value.{AbsoluteContractId, VersionedValue}
 import com.daml.lf.value.ValueCoder
 import com.daml.lf.value.ValueCoder.{DecodeError, EncodeError}
-import com.daml.ledger.EventId
 
 trait TransactionSerializer {
